Revert "[GH-10800] convert bot account to user account (#10803)"
This reverts commit 2130e9f0b1b7f91300b0effa26c8e35d7c821523.

+++ store/sqlstore/bot_store.go
@@ -233,6 +233,12 @@ func (us SqlBotStore) Update(bot *model.Bot) store.StoreChannel {
 // If the corresponding user is to be deleted, it must be done via the user store.
 func (us SqlBotStore) PermanentDelete(botUserId string) store.StoreChannel {
 	return store.Do(func(result *store.StoreResult) {
+		userResult := <-us.User().PermanentDelete(botUserId)
+		if userResult.Err != nil {
+			result.Err = userResult.Err
+			return
+		}
+
 		if _, err := us.GetMaster().Exec(`
 			DELETE FROM
 			    Bots
